// src/http-busca-search/models/modelsButton.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */

export interface ModelsButton {
  descricao?: string
  is_enabled?: boolean
  ordem?: number
  titulo?: string
  url_service?: string
}