// src/http-pref-rio-carta-servicos/models/channel.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */
import type { ChannelChannelType } from './channelChannelType'
import type { ChannelType } from './channelType'

export interface Channel {
  order?: number
  type?: ChannelType
  endereco?: string
  titulo?: string
  /** Gerado automaticamente para canais Presenciais com endereço */
  mapUrl?: string
  value?: string
  channelType?: ChannelChannelType
  /** Gerado automaticamente para canais WhatsApp */
  whatsappLink?: string
}