// src/http-pref-rio-carta-servicos/models/serviceFilteredItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */
import type { ServiceFilteredItemArticleStatus } from './serviceFilteredItemArticleStatus'

/**
 * Versão resumida usada na resposta de /themes?filter_category=
 */
export interface ServiceFilteredItem {
  slug?: string
  name?: string
  summary?: string
  serviceCatalogId?: string
  articleId?: string
  articleVersion?: number
  articleStatus?: ServiceFilteredItemArticleStatus
  createdDate?: string
  lastModifiedDate?: string
  lastPublishedDate?: string
  responsibleOrgUnit?: string
  responsibleOrgUnitShortName?: string
  serviceDeadline?: string
  /** @nullable */
  cost?: string | null
  /** @nullable */
  costValue?: number | null
  /** @nullable */
  costNotes?: string | null
  isFree?: boolean
}