// src/http-pref-rio-carta-servicos/models/serviceListItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */
import type { ServiceListItemArticleStatus } from './serviceListItemArticleStatus'

export interface ServiceListItem {
  slug?: string
  name?: string
  serviceCatalogId?: string
  articleId?: string
  articleVersion?: number
  articleStatus?: ServiceListItemArticleStatus
  createdDate?: string
  lastModifiedDate?: string
  lastPublishedDate?: string
  responsibleOrgUnit?: string
  responsibleOrgUnitShortName?: string
  summary?: string
  serviceDeadline?: string
  /** @nullable */
  serviceDeadlineNotes?: string | null
  /** @nullable */
  cost?: string | null
  /** @nullable */
  costValue?: number | null
  /** @nullable */
  costNotes?: string | null
  isFree?: boolean
  fixarDestaque?: boolean
  allowTicketSubmission?: boolean
  allowsAnonymity?: boolean
}